+++ src/filament_manager/__init__.py
@@ -1,3 +1,3 @@
 """Filament Manager application package."""
 
-__version__ = "0.5.6"
+__version__ = "0.5.7"